+++ NEWS.md
@@ -1,5 +1,7 @@
 # dplyr 0.4.3.9000
 
+* `do.data.table()` and `do.tbl_dt()` now work (#1081).
+
 * `top_n()` can now also select bottom `n` values by passing a negative value
   to `n` (#1008, #1352).